moved bot.py into examples
